+++ src/hrdmc/estimators/__init__.py
@@ -1,16 +1,24 @@
+from hrdmc.estimators.cloud import CloudMomentResult, estimate_cloud_moments
 from hrdmc.estimators.density import (
+    density_support_edges,
     estimate_density_profile,
     estimate_open_line_density_profile,
     integrate_density_profile,
 )
 from hrdmc.estimators.local_energy import LocalEnergyResult, estimate_local_energy
 from hrdmc.estimators.pair_distribution import PairDistributionResult, estimate_pair_distribution
-from hrdmc.estimators.structure_factor import StructureFactorResult, estimate_static_structure_factor
+from hrdmc.estimators.structure_factor import (
+    StructureFactorResult,
+    estimate_static_structure_factor,
+)
 
 __all__ = [
     "LocalEnergyResult",
+    "CloudMomentResult",
     "PairDistributionResult",
     "StructureFactorResult",
+    "density_support_edges",
+    "estimate_cloud_moments",
     "estimate_density_profile",
     "estimate_open_line_density_profile",
     "estimate_local_energy",