+++ src/hrdmc/theory/__init__.py
@@ -6,7 +6,14 @@ from hrdmc.theory.hard_rods import (
     hard_rod_finite_ring_energy_per_particle,
     invert_hard_rod_chemical_potential,
 )
-from hrdmc.theory.lda import LDADensityProfile, lda_density_profile, lda_total_energy
+from hrdmc.theory.lda import (
+    LDADensityProfile,
+    lda_density_profile,
+    lda_mean_square_radius,
+    lda_rms_radius,
+    lda_support_edges,
+    lda_total_energy,
+)
 
 __all__ = [
     "LDADensityProfile",
@@ -17,5 +24,8 @@ __all__ = [
     "hard_rod_finite_ring_energy_per_particle",
     "invert_hard_rod_chemical_potential",
     "lda_density_profile",
+    "lda_mean_square_radius",
+    "lda_rms_radius",
+    "lda_support_edges",
     "lda_total_energy",
 ]